+++ MRML/vtkMRMLROS2PublisherDefaultNodes.h
@@ -7,14 +7,14 @@
 VTK_MRML_ROS_PUBLISHER_NATIVE_H(std::string, String);
 VTK_MRML_ROS_PUBLISHER_NATIVE_H(bool, Bool);
 VTK_MRML_ROS_PUBLISHER_NATIVE_H(int, Int);
-VTK_MRML_ROS_PUBLISHER_NATIVE_H(double, Float);
+VTK_MRML_ROS_PUBLISHER_NATIVE_H(double, Double);
 
 #include <vtkMatrix4x4.h>
 #include <vtkTransformCollection.h>
 #include <vtkDenseArray.h>
 
 VTK_MRML_ROS_PUBLISHER_VTK_H(vtkIntArray, IntArray);
-VTK_MRML_ROS_PUBLISHER_VTK_H(vtkDoubleArray, FloatArray);
+VTK_MRML_ROS_PUBLISHER_VTK_H(vtkDoubleArray, DoubleArray);
 VTK_MRML_ROS_PUBLISHER_VTK_H(vtkDenseArray<int>, IntNArray);
 VTK_MRML_ROS_PUBLISHER_VTK_H(vtkMatrix4x4, PoseStamped);
 VTK_MRML_ROS_PUBLISHER_VTK_H(vtkDoubleArray, WrenchStamped);
